Build TFM & MCUBoot as a subproject
This also links the App at the offset expected by TFM and uses a
non-secure address in the startup code

This allows for firmware update, MCUBoot sytle.

The modifications include:
 * Reserve room in flash for the MCUBoot header (0x400)
 * Sign firmware parts as part of the build process

The total size of RAM in MPS3 AN524 is 128KB out of which 96KB is
allocated to TF-M. Update app linker file to allocate remaining RAM. In
addition reduce stack size from `0x1800` to `0x1000` to keep heap size
at `0x1000`.

Finally, we call some PSA APIs in the example application to test this
integration.

Signed-off-by: Jimmy Brisson <jimmy.brisson@linaro.org>
Signed-off-by: Devaraj Ranganna <devaraj.ranganna@linaro.org>